% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsaServer}
\alias{bsaServer}
\title{Análisis de significación biológica (Server)}
\usage{
bsaServer(id, data_type, data, org, pvcoff, padmethod, struct)
}
\arguments{
\item{id}{Identificador.}
}
\description{
Análisis de significación biológica (Server)
}
