% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combina}
\alias{combina}
\title{Combinatoria}
\usage{
combina(x)
}
\arguments{
\item{x}{Vector de caracteres.}
}
\value{
Vector con la combinación de elementos de \code{x}.
}
\description{
Combinatoria
}
\examples{
vector <- c("A", "B", "C")
combina(vector)
}
