% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combina_int}
\alias{combina_int}
\title{Combinatoria 2}
\usage{
combina_int(x)
}
\arguments{
\item{x}{Vector de caracteres.}
}
\value{
Vector con la combinación de elementos de \code{x}. Cada elemento de
\code{x} estará introducido en un paréntesis.
}
\description{
Combinatoria 2
}
\examples{
vector <- c("A-B", "A-C", "B-C")
combina_int(vector)
}
