% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{deaServer}
\alias{deaServer}
\title{Análisis de expresión diferencial (Server)}
\usage{
deaServer(id, norm_data, data_type, pack)
}
\arguments{
\item{id}{Identificador.}

\item{norm_data}{Datos normalizados en uno de los módulos anteriores.}

\item{data_type}{Tipo de datos, seleccionado en la aplicación inicial.}

\item{pack}{Paquete de anotaciones introducido en el módulo
\code{norm_microarray}.}
}
\description{
Análisis de expresión diferencial (Server)
}
