% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_MetabMC.R
\name{norm_MetabMCServer}
\alias{norm_MetabMCServer}
\title{Normalización de datos de concentraciones de metabolitos (Server)}
\usage{
norm_MetabMCServer(id, data)
}
\arguments{
\item{id}{Identificador.}

\item{data}{Datos cargados en el módulo \code{read_data}.}
}
\description{
Normalización de datos de concentraciones de metabolitos (Server)
}
