% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_MetabMC.R
\name{norm_MetabMCUI}
\alias{norm_MetabMCUI}
\title{Normalización de datos de concentraciones de metabolitos (UI)}
\usage{
norm_MetabMCUI(id)
}
\arguments{
\item{id}{Identificador.}
}
\description{
Normalización de datos de concentraciones de metabolitos (UI)
}
