% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_MetabRS.R
\name{norm_MetabRSServer}
\alias{norm_MetabRSServer}
\title{Preprocesamiento y normalización de datos de espectros brutos de GC/LC-MS (server)}
\usage{
norm_MetabRSServer(id, data)
}
\arguments{
\item{id}{Identificador.}

\item{data}{Datos cargados en el módulo \code{read_data}.}
}
\description{
Preprocesamiento y normalización de datos de espectros brutos de GC/LC-MS (server)
}
