% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_MetabSB.R
\name{norm_MetabSBServer}
\alias{norm_MetabSBServer}
\title{Preprocesamiento y normalización de datos de contenedores de espectros de MS (Server)}
\usage{
norm_MetabSBServer(id, data)
}
\arguments{
\item{id}{Identificador.}

\item{data}{Datos cargados en el módulo \code{read_data}.}
}
\description{
Preprocesamiento y normalización de datos de contenedores de espectros de MS (Server)
}
