% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_microarray.R
\name{norm_microarrayServer}
\alias{norm_microarrayServer}
\title{Normalización de datos de microarray (Server)}
\usage{
norm_microarrayServer(id, data)
}
\arguments{
\item{id}{Identificador.}

\item{data}{Datos cargados en el módulo \code{read_data}.}
}
\description{
Normalización de datos de microarray (Server)
}
