% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_RNASeq.R
\name{norm_RNASeqServer}
\alias{norm_RNASeqServer}
\title{Normalización de datos de RNA-Seq (Server)}
\usage{
norm_RNASeqServer(id, data)
}
\arguments{
\item{id}{Identificador.}

\item{data}{Datos cargados en el módulo \code{read_data}.}
}
\description{
Normalización de datos de RNA-Seq (Server)
}
