% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_dataServer}
\alias{read_dataServer}
\title{Lectura de datos (Server)}
\usage{
read_dataServer(id, data_type)
}
\arguments{
\item{id}{Identificador.}

\item{data_type}{Tipo de datos, seleccionado en la aplicación inicial.}
}
\description{
Lectura de datos (Server)
}
