% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{TFMjrufvApp}
\alias{TFMjrufvApp}
\title{Framework para el análisis de datos ómicos}
\usage{
TFMjrufvApp()
}
\description{
Permite realizar un análisis completo de diferentes tipos de datos ómicos a
través de diferentes pestañas correspondientes a los pasos a dar para cargar
los datos, preprocesarlos y normalizarlos, realizar análisis estadísticos,
anotar los resultados y obtener análisis de significación biológica.
}
